# Improve `F.resize_images` speed