--- a/chainer/functions/array/resize_images.py
+++ b/chainer/functions/array/resize_images.py
@@ -6,6 +6,169 @@ from chainer import function_node
 from chainer.utils import type_check
 
 
+def _infer_lines(B, C, H, W, out_H, out_W, kH, kW):
+    target_size = 2 ** 17
+    line_size = B * C * (H * W // out_H + kH * kW * out_W)
+    target_lines = target_size // line_size
+
+    if target_lines < out_H:
+        lines = 1
+        while True:
+            next_lines = lines * 2
+            if next_lines > target_lines:
+                break
+            lines = next_lines
+    else:
+        lines = out_H
+
+    return lines
+
+
+def interpolate_bilinear_cpu(x, v, u, vw, uw):
+    B, C, H, W = x.shape
+    out_H, out_W = v.shape
+
+    # Interpolation is done by each output panel (i.e. multi lines)
+    # in order to better utilize CPU cache memory.
+    lines = _infer_lines(B, C, H, W, out_H, out_W, 2, 2)
+
+    vcol = numpy.empty((2, lines, out_W), dtype=v.dtype)
+    ucol = numpy.empty((2, lines, out_W), dtype=u.dtype)
+    wcol = numpy.empty((2, 2, lines, out_W), dtype=x.dtype)
+
+    y = numpy.empty((B * C, out_H * out_W), dtype=x.dtype)
+
+    for i in range(0, out_H, lines):
+        l = min(lines, out_H - i)
+        vcol = vcol[:, :l]
+        ucol = ucol[:, :l]
+        wcol = wcol[:, :, :l]
+        i_end = i + l
+
+        # indices
+        vcol[0] = v[i:i_end]
+        ucol[0] = u[i:i_end]
+        numpy.add(vcol[0], 1, out=vcol[1])
+        numpy.add(ucol[0], 1, out=ucol[1])
+        numpy.minimum(vcol[1], H - 1, out=vcol[1])
+        numpy.minimum(ucol[1], W - 1, out=ucol[1])
+
+        # weights
+        #   wcol[0, 0] = (1 - uw) * (1 - vw)
+        #   wcol[0, 1] = uw * (1 - vw)
+        #   wcol[1, 0] = (1 - uw) * vw
+        #   wcol[1, 1] = uw * vw
+        wcol[0, 1] = uw[i:i_end]
+        numpy.subtract(1, wcol[0, 1], out=wcol[0, 0])
+        numpy.multiply(wcol[0], vw[i:i_end], out=wcol[1])
+        wcol[0] -= wcol[1]
+
+        # packing to the panel whose shape is (B, C, 2, 2, l, out_W)
+        panel = x[:, :, vcol[:, None], ucol[None, :]]
+
+        # interpolation
+        panel = panel.reshape((B * C, 4, l * out_W))
+        weights = wcol.reshape((4, l * out_W))
+        iout = i * out_W
+        iout_end = i_end * out_W
+        numpy.einsum('ijk,jk->ik', panel, weights, out=y[:, iout:iout_end])
+        del panel, weights
+
+    return y.reshape((B, C, out_H, out_W))
+
+
+def interpolate_bilinear_gpu(x, v, u, vw, uw):
+    B, C, H, W = x.shape
+    out_H, out_W = v.shape
+    y = cuda.cupy.empty((B, C, out_H, out_W), dtype=x.dtype)
+
+    cuda.elementwise(
+        'raw T x, S v, S u, T vw, T uw, S H, S W, S outsize', 'T y', '''
+        // indices
+        S v0 = v;
+        S v1 = min(v + 1, (S)(H - 1));
+        S u0 = u;
+        S u1 = min(u + 1, (S)(W - 1));
+        // weights
+        T w0 = (1 - vw) * (1 - uw);
+        T w1 = (1 - vw) * uw;
+        T w2 = vw * (1 - uw);
+        T w3 = vw * uw;
+        // fetch
+        S offset = i / outsize * H * W;
+        T px0 = x[offset + v0 * W + u0];
+        T px1 = x[offset + v0 * W + u1];
+        T px2 = x[offset + v1 * W + u0];
+        T px3 = x[offset + v1 * W + u1];
+        // interpolate
+        y = (w0 * px0 + w1 * px1) + (w2 * px2 + w3 * px3);
+        ''', 'resize_images_interpolate_bilinear'
+    )(x, v, u, vw, uw, H, W, out_H * out_W, y)
+    return y
+
+
+def interpolate_grad_bilinear_cpu(gy, v, u, vw, uw, H, W):
+    B, C, out_H, out_W = gy.shape
+
+    # indices
+    vcol = numpy.empty((2, out_H, out_W), dtype=v.dtype)
+    ucol = numpy.empty((2, out_H, out_W), dtype=u.dtype)
+    vcol[0] = v
+    ucol[0] = u
+    numpy.add(vcol[0], 1, out=vcol[1])
+    numpy.add(ucol[0], 1, out=ucol[1])
+    numpy.minimum(vcol[1], H - 1, out=vcol[1])
+    numpy.minimum(ucol[1], W - 1, out=ucol[1])
+
+    # weights
+    wcol = numpy.empty((2, 2, out_H, out_W), dtype=gy.dtype)
+    wcol[0, 1] = uw
+    numpy.subtract(1, wcol[0, 1], out=wcol[0, 0])
+    numpy.multiply(wcol[0], vw, out=wcol[1])
+    wcol[0] -= wcol[1]
+
+    # grad
+    gycol = gy.reshape((B * C, 1, 1, out_H, out_W)) * wcol
+
+    # ravel everything and use `bincount`
+    indices = (vcol[:, None] * W + ucol[None, :]).ravel()
+    offsets = numpy.arange(0, B * C * H * W, H * W, dtype=v.dtype)
+    indices = (offsets[:, None] + indices).ravel()
+    gx = numpy.bincount(indices, weights=gycol.ravel(),
+                        minlength=(B * C * H * W))
+    gx = gx.astype(gy.dtype, copy=False)
+
+    return gx.reshape((B, C, H, W))
+
+
+def interpolate_grad_bilinear_gpu(gy, v, u, vw, uw, H, W):
+    B, C, out_H, out_W = gy.shape
+    gx = cuda.cupy.zeros((B * C, H, W), dtype=gy.dtype)
+
+    cuda.elementwise(
+        'T gy, S v, S u, T vw, T uw, S H, S W, S outsize', 'raw T gx', '''
+        // indices
+        S v0 = v;
+        S v1 = min(v + 1, (S)(H - 1));
+        S u0 = u;
+        S u1 = min(u + 1, (S)(W - 1));
+        // weights
+        T w0 = (1 - vw) * (1 - uw);
+        T w1 = (1 - vw) * uw;
+        T w2 = vw * (1 - uw);
+        T w3 = vw * uw;
+        // scatter
+        S offset = i / outsize * H * W;
+        atomicAdd(&gx[offset + v0 * W + u0], w0 * gy);
+        atomicAdd(&gx[offset + v0 * W + u1], w1 * gy);
+        atomicAdd(&gx[offset + v1 * W + u0], w2 * gy);
+        atomicAdd(&gx[offset + v1 * W + u1], w3 * gy);
+        ''', 'resize_images_interpolate_grad_bilinear'
+    )(gy, v, u, vw, uw, H, W, out_H * out_W, gx)
+
+    return gx.reshape((B, C, H, W))
+
+
 class ResizeImages(function_node.FunctionNode):
 
     def __init__(self, output_shape):
@@ -25,38 +188,30 @@ class ResizeImages(function_node.FunctionNode):
         x, = inputs
         xp = backend.get_array_module(x)
 
-        B, C, H, W = x.shape
+        _, C, H, W = x.shape
+        out_H, out_W = self.out_H, self.out_W
 
-        u_1d = xp.linspace(0, W - 1, num=self.out_W)
-        v_1d = xp.linspace(0, H - 1, num=self.out_H)
-        grid = xp.meshgrid(u_1d, v_1d)
-        # u, v are of shape (out_H * out_W,)
-        u = grid[0].ravel()
-        v = grid[1].ravel()
+        # Compute indices and weights.
+        v = xp.linspace(0, H - 1, num=out_H, dtype=numpy.float)
+        u = xp.linspace(0, W - 1, num=out_W, dtype=numpy.float)
+        vw, v = xp.modf(v)
+        uw, u = xp.modf(u)
+        v = v.astype(numpy.intp)
+        u = u.astype(numpy.intp)
+        vw = vw.astype(x.dtype)
+        uw = uw.astype(x.dtype)
 
-        # indices of the 2x2 pixel neighborhood surrounding the coordinates
-        u0 = xp.floor(u).astype(numpy.int32)
-        u0 = u0.clip(0, W - 2)
-        u1 = u0 + 1
-        v0 = xp.floor(v).astype(numpy.int32)
-        v0 = v0.clip(0, H - 2)
-        v1 = v0 + 1
+        # Meshgrid-like operation. Meshgrid can cause
+        # performance loss due to memory consumption.
+        v = xp.broadcast_to(v[:, None], (out_H, out_W))
+        u = xp.broadcast_to(u[None, :], (out_H, out_W))
+        vw = xp.broadcast_to(vw[:, None], (out_H, out_W))
+        uw = xp.broadcast_to(uw[None, :], (out_H, out_W))
 
-        # weights
-        w1 = (u1 - u) * (v1 - v)
-        w2 = (u - u0) * (v1 - v)
-        w3 = (u1 - u) * (v - v0)
-        w4 = (u - u0) * (v - v0)
-        w1 = w1.astype(x.dtype, copy=False)
-        w2 = w2.astype(x.dtype, copy=False)
-        w3 = w3.astype(x.dtype, copy=False)
-        w4 = w4.astype(x.dtype, copy=False)
-
-        y = (w1[None, None, :] * x[:, :, v0, u0] +
-             w2[None, None, :] * x[:, :, v0, u1] +
-             w3[None, None, :] * x[:, :, v1, u0] +
-             w4[None, None, :] * x[:, :, v1, u1])
-        y = y.reshape(B, C, self.out_H, self.out_W)
+        if xp is numpy:
+            y = interpolate_bilinear_cpu(x, v, u, vw, uw)
+        else:
+            y = interpolate_bilinear_gpu(x, v, u, vw, uw)
         return y,
 
     def backward(self, indexes, grad_outputs):
@@ -81,48 +236,33 @@ class ResizeImagesGrad(function_node.FunctionNode):
         )
 
     def forward(self, inputs):
-        xp = backend.get_array_module(*inputs)
         gy, = inputs
+        xp = backend.get_array_module(gy)
 
-        B, C, H, W = self.input_shape
+        _, C, H, W = self.input_shape
+        out_H, out_W = self.out_H, self.out_W
 
-        u_1d = xp.linspace(0, W - 1, num=self.out_W)
-        v_1d = xp.linspace(0, H - 1, num=self.out_H)
-        grid = xp.meshgrid(u_1d, v_1d)
-        # u, v are of shape (out_H * out_W,)
-        u = grid[0].ravel()
-        v = grid[1].ravel()
+        # Compute indices and weights.
+        v = xp.linspace(0, H - 1, num=out_H, dtype=numpy.float)
+        u = xp.linspace(0, W - 1, num=out_W, dtype=numpy.float)
+        vw, v = xp.modf(v)
+        uw, u = xp.modf(u)
+        v = v.astype(numpy.intp)
+        u = u.astype(numpy.intp)
+        vw = vw.astype(gy.dtype)
+        uw = uw.astype(gy.dtype)
 
-        # indices of the 2x2 pixel neighborhood surrounding the coordinates
-        u0 = xp.floor(u).astype(numpy.int32)
-        u0 = u0.clip(0, W - 2)
-        u1 = u0 + 1
-        v0 = xp.floor(v).astype(numpy.int32)
-        v0 = v0.clip(0, H - 2)
-        v1 = v0 + 1
+        # Meshgrid-like operation. Meshgrid can cause
+        # performance loss due to memory consumption.
+        v = xp.broadcast_to(v[:, None], (out_H, out_W))
+        u = xp.broadcast_to(u[None, :], (out_H, out_W))
+        vw = xp.broadcast_to(vw[:, None], (out_H, out_W))
+        uw = xp.broadcast_to(uw[None, :], (out_H, out_W))
 
-        # weights
-        wu0 = u - u0
-        wu1 = u1 - u
-        wv0 = v - v0
-        wv1 = v1 - v
-        wu0 = wu0.astype(gy.dtype, copy=False)
-        wu1 = wu1.astype(gy.dtype, copy=False)
-        wv0 = wv0.astype(gy.dtype, copy=False)
-        wv1 = wv1.astype(gy.dtype, copy=False)
-
-        # --- gx
         if xp is numpy:
-            scatter_add = numpy.add.at
+            gx = interpolate_grad_bilinear_cpu(gy, v, u, vw, uw, H, W)
         else:
-            scatter_add = cuda.cupyx.scatter_add
-
-        gx = xp.zeros(self.input_shape, dtype=gy.dtype)
-        gy = gy.reshape(B, C, -1)
-        scatter_add(gx, (slice(None), slice(None), v0, u0), gy * wu1 * wv1)
-        scatter_add(gx, (slice(None), slice(None), v0, u1), gy * wu0 * wv1)
-        scatter_add(gx, (slice(None), slice(None), v1, u0), gy * wu1 * wv0)
-        scatter_add(gx, (slice(None), slice(None), v1, u1), gy * wu0 * wv0)
+            gx = interpolate_grad_bilinear_gpu(gy, v, u, vw, uw, H, W)
         return gx,
 
     def backward(self, indexes, grad_outputs):