@grafi-tt 2018-12-18
C++ is more suitable to write such a code, so if developers prefers ChainerX for performance-sensitive operations, I'll withdraw this PR.

I guess tight loop written in C++ outperforms any pure NumPy implementation, even without SIMD. But benchmark will be necessary.