@grafi-tt 2019-03-05
@okuta @mitmul Sorry for late reply. Now `broadcast_arrays` is used instead of `broadcast_to` method.